+++ Logging/src/main/java/runday/domain/RunningStarted.java
@@ -12,6 +12,6 @@ public class RunningStarted extends AbstractEvent {
     private Long id;
     private String userId;
     private Date startTime;
-    private Long startLatitude;
-    private Long startLongitude;
+    private Float startLatitude;
+    private Float startLongitude;
 }
